August 16, 2017
Checked build for newer versions of:
 - MaxOS 10.7 - 10.11 (note1)
 - XCode for those OS except 8.x
 - Qt 4.8.7 and newer (note2)
 - OpenCV 2.4.13 and newer (note3)
 - DeckLink SDK 9.x - 10.9.x

Note1: Currently SCCToolKit does not work on macOS 10.12 (Sierra)
because QTKit was deprecated. We are currently working to change to
AVFoundation to work on 10.12 and coming 10.13.
Note2: There is a long term support (LTS) version in 5.6. We recommend
using it.
Note3: We recommend using 3.1 or newer.

Bugs solved:
 - FindDeckLink module now works with system environment variables.
 - Old C structures have been replaced to C++ classes in Examples.
 - Fixed for OpenGL support changes in Qt 5.x.